requirements.txt
google-generativeai
sentence-transformers
faiss-cpu
python-dotenv
pandas
tqdm
transformers
peft
trl
bitsandbytes